chore: state 클래스를 뷰모델로 옮깁니다
